% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_abundance_transition.R
\name{calcecdf}
\alias{calcecdf}
\title{Calculate ECDF given a vector of counts}
\usage{
calcecdf(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
the empirical CDF representing the rankings in x
}
\description{
Calculate ECDF given a vector of counts
}
