% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_abundance_transition.R
\name{calcval}
\alias{calcval}
\title{Map denoised quantiles back to counts}
\usage{
calcval(Q_denoised, Q_observed, values)
}
\arguments{
\item{Q_denoised}{quantiles after denoising}

\item{Q_observed}{the quantiles of the original observed values}

\item{values}{the original observed values}
}
\value{
the interpolated denoised values
}
\description{
Map denoised quantiles back to counts
}
