% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_abundance_transition.R
\name{count2ecdf}
\alias{count2ecdf}
\title{Calculate ECDF for every column of a count matrix}
\usage{
count2ecdf(X)
}
\arguments{
\item{X}{a matrix of values. rows are samples and columns are features}
}
\description{
Calculate ECDF for every column of a count matrix
}
