% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{cv.deconv}
\alias{cv.deconv}
\title{Cross validation of deconvolution}
\usage{
cv.deconv(y, X, lambda = 0.1, cv.fold = 5, type = c("l1", "l2"))
}
\arguments{
\item{y}{response vector}

\item{X}{predictor matrix}

\item{lambda}{penalty parameter}

\item{cv.fold}{number of folds for repeated fitting}

\item{type}{choose between absolute loss or square loss when fitting deconvolution}
}
\value{
weights, mean error for the and chosen penalty parameter
}
\description{
Cross validation of deconvolution
}
