% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{deconv.l1}
\alias{deconv.l1}
\title{Deconvolve ecdfs of one sample with others (l1 loss)}
\usage{
deconv.l1(y, X, lambda)
}
\arguments{
\item{y}{response vector}

\item{X}{predictor matrix}

\item{lambda}{penalty parameter}
}
\value{
weights, denoised value and chosen penalty parameter
}
\description{
Deconvolve ecdfs of one sample with others (l1 loss)
}
