% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{deconvolve every row of a matrix with ECDF values}
\usage{
denoise(Q, lambda = 0.1, cv.fold = 10, type = c("l1", "l2"), ncores = 1)
}
\arguments{
\item{Q}{input matrix whose entries are ECDF values}

\item{lambda}{penalty parameter}

\item{cv.fold}{number of folds for cross validation}

\item{type}{choose between absolute loss or square loss when fitting deconvolution}

\item{ncores}{number of cores for parallel processing, by default 1}
}
\value{
denoised matrix
}
\description{
deconvolve every row of a matrix with ECDF values
}
