% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREC.R
\name{drec}
\alias{drec}
\title{Denoise metagenomics count table}
\usage{
drec(X, lambdas = c(0.1, 0.5, 1, 5, 10, 20, 50, 100), ncores = 1, cv.fold = 10)
}
\arguments{
\item{X}{raw count matrix, rows are samples and columns are features}

\item{lambdas}{penalty parameters to choose from}

\item{ncores}{number of cores for parallel computing}

\item{cv.fold}{number of folds for cross validation}
}
\value{
denoised count matrix
}
\description{
Denoise metagenomics count table
}
