% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_abundance_transition.R
\name{ecdf2count}
\alias{ecdf2count}
\title{Map denoised quantiles back to counts for each column}
\usage{
ecdf2count(Q_denoised, Q_observed, count_observed)
}
\arguments{
\item{Q_denoised}{denoised empirical CDF table}

\item{Q_observed}{original empirical CDF table}

\item{count_observed}{observed counts table}
}
\value{
denoised counts
}
\description{
Map denoised quantiles back to counts for each column
}
