% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{cv.deconv}
\alias{cv.deconv}
\title{Cross validation of deconvolution}
\usage{
cv.deconv(y, X, lambdas = c(0.1, 0.5, 1, 5, 10, 20, 50, 100), cv.fold = 5)
}
\arguments{
\item{y}{response vector}

\item{X}{predictor matrix}

\item{lambdas}{penalty parameters to try}

\item{cv.fold}{number of folds for cross validation}
}
\value{
weights, mean error for the and chosen penalty parameter
}
\description{
Cross validation of deconvolution
}
