% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{cv.deconv}
\alias{cv.deconv}
\title{Cross validation of deconvolution}
\usage{
cv.deconv(y, X, lambdas = c(0.001, 0.01, 0.1))
}
\arguments{
\item{y}{response vector}

\item{X}{predictor matrix}

\item{lambdas}{penalty parameters to try}
}
\value{
weights, mean error for the and chosen penalty parameter
}
\description{
Cross validation of deconvolution
}
