% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{deconv}
\alias{deconv}
\title{Deconvolve ecdfs of one sample with others}
\usage{
deconv(y, X, lambdas = c(0.001, 0.01, 0.1))
}
\arguments{
\item{y}{response vector}

\item{X}{predictor matrix}

\item{lambdas}{penalty parameters to choose from}
}
\value{
weights, denoised value and chosen penalty parameter
}
\description{
Deconvolve ecdfs of one sample with others
}
