% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{deconvolve every row of a matrix with ECDF values}
\usage{
denoise(
  Q,
  lambdas = c(0.1, 0.5, 1, 5, 10, 20, 50, 100),
  ncores = 1,
  cv.fold = 10
)
}
\arguments{
\item{Q}{input matrix whose entries are ECDF values}

\item{lambdas}{penalty parameters to choose from}

\item{ncores}{number of cores for parallel processing, by default 1}

\item{cv.fold}{number of folds for cross validation}
}
\value{
denoised matrix
}
\description{
deconvolve every row of a matrix with ECDF values
}
