% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{deconvolve every row of a matrix with ECDF values}
\usage{
denoise(Q, lambdas = c(0.001, 0.01, 0.1, 1, 10, 100), ncores = 1)
}
\arguments{
\item{Q}{input matrix whose entries are ECDF values}

\item{lambdas}{penalty parameters to choose from}

\item{ncores}{number of cores for parallel processing, by default 1}
}
\value{
denoised matrix
}
\description{
deconvolve every row of a matrix with ECDF values
}
