% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREC.R
\name{drec}
\alias{drec}
\title{Denoise metagenomics count table}
\usage{
drec(X, lambdas = c(0.001, 0.01, 0.1), ncores = 1)
}
\arguments{
\item{X}{raw count matrix, rows are samples and columns are features}

\item{lambdas}{penalty parameters to choose from}

\item{ncores}{number of cores for parallel computing}
}
\value{
denoised count matrix
}
\description{
Denoise metagenomics count table
}
