% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREC.R
\name{drec}
\alias{drec}
\title{Denoise metagenomics count table}
\usage{
drec(X, lambda = 0.1, cv.fold = 10, type = c("l1", "l2"), ncores = 1)
}
\arguments{
\item{X}{raw count matrix, rows are samples and columns are features}

\item{lambda}{penalty parameter}

\item{cv.fold}{number of folds for cross validation}

\item{type}{l1 (absolute loss) or l2 (quadratic loss)}

\item{ncores}{number of cores for parallel computing}
}
\value{
denoised count matrix
}
\description{
Denoise metagenomics count table
}
